teste 4

+++ pages/aluno/aluno_excluir.php
@@ -1,7 +1,8 @@
 <?php
+include "../../includes/sessao.php";
 include "../../includes/conexao.php";
-$id = $_GET['id'];
-
-$conn->query("DELETE FROM alunos WHERE id=$id");
-
-echo "<script>alert('Excluído com sucesso!'); location.href='aluno_lista.php';</script>";
+$id = (int)($_GET['id'] ?? 0);
+if ($id) {
+    $conn->query("DELETE FROM alunos WHERE id=$id");
+}
+echo "<script>location.href='aluno_lista.php';</script>";
