+++ pages/pagamentos/pagamentos_excluir.php
@@ -1,7 +1,6 @@
 <?php
+include "../../includes/sessao.php";
 include "../../includes/conexao.php";
-
-$id = (int)$_GET['id'];
-$conn->query("DELETE FROM pagamentos WHERE id=$id");
-
-echo "<script>alert('Pagamento excluído'); location.href='pagamentos_lista.php';</script>";
+$id = (int)($_GET['id'] ?? 0);
+if ($id) { $conn->query("DELETE FROM pagamentos WHERE id=$id"); }
+echo "<script>location.href='pagamentos_lista.php';</script>";
